refactor: implement Phase 6 - State Layer
- Create src/state/ directory for global state management
- Move authStore from stores/ to state/ with enhanced structure:
  - Separate AuthState and AuthActions types
  - Add selectors (selectUser, selectIsAuthenticated, selectIsLoading, etc.)
  - Add clearError action
  - Add selectIsAdmin selector

- Update all imports across the codebase:
  - App.tsx, Login.tsx, Settings.tsx, MainLayout.tsx
  - authStore.test.ts

- Remove legacy src/stores/ folder

All 194 tests passing, build successful, lint clean.
