+++ admin/src/pages/Login.tsx
@@ -2,7 +2,7 @@
 import { Card, Button, Space, Typography, message } from 'antd';
 import { GoogleOutlined, FolderOutlined } from '@ant-design/icons';
 import { useNavigate } from 'react-router-dom';
-import { useAuthStore } from '../stores/authStore';
+import { useAuthStore } from '../state';
 import './Login.css';
 
 const { Title, Text } = Typography;
