+++ admin/src/state/authStore.ts
@@ -1,4 +1,7 @@
-// 인증 상태 관리 (Zustand + Firebase)
+/**
+ * Auth State Store - 인증 상태 관리 (Zustand)
+ * State Layer: 전역 상태 관리
+ */
 import { create } from 'zustand';
 import type { User } from '../core/types';
 import {
@@ -8,20 +11,31 @@ import {
   onAuthChange,
 } from '../data/repository';
 
+// ============ Types ============
+
 interface AuthState {
   user: User | null;
   isAuthenticated: boolean;
   isLoading: boolean;
   error: string | null;
+}
+
+interface AuthActions {
   login: () => Promise<void>;
   logout: () => Promise<void>;
   initializeAuth: () => Promise<void>;
+  clearError: () => void;
 }
 
-export const useAuthStore = create<AuthState>((set) => ({
+type AuthStore = AuthState & AuthActions;
+
+// ============ Store ============
+
+export const useAuthStore = create<AuthStore>((set) => ({
+  // Initial state
   user: null,
   isAuthenticated: false,
-  isLoading: true, // 초기 로딩 상태
+  isLoading: true,
   error: null,
 
   // Google 로그인
@@ -73,9 +87,46 @@ export const useAuthStore = create<AuthState>((set) => ({
       set({ error: errorMessage, isLoading: false });
     }
   },
+
+  // 에러 클리어
+  clearError: () => {
+    set({ error: null });
+  },
 }));
 
-// 개발용 목 로그인 함수 (Firebase 연결 전 테스트용)
+// ============ Selectors ============
+
+/**
+ * 사용자 정보만 선택
+ */
+export const selectUser = (state: AuthStore) => state.user;
+
+/**
+ * 인증 여부만 선택
+ */
+export const selectIsAuthenticated = (state: AuthStore) => state.isAuthenticated;
+
+/**
+ * 로딩 상태만 선택
+ */
+export const selectIsLoading = (state: AuthStore) => state.isLoading;
+
+/**
+ * 에러 상태만 선택
+ */
+export const selectError = (state: AuthStore) => state.error;
+
+/**
+ * 관리자 여부 선택
+ */
+export const selectIsAdmin = (state: AuthStore) => state.user?.role === 'admin';
+
+// ============ Dev Utils ============
+
+/**
+ * 개발용 목 로그인 함수 (Firebase 연결 전 테스트용)
+ * @deprecated 개발/테스트 용도로만 사용
+ */
 export const mockLogin = async (): Promise<User> => {
   return new Promise((resolve) => {
     setTimeout(() => {
