+++ admin/src/state/index.ts
@@ -1,0 +1,15 @@
+/**
+ * State Layer - 전역 상태 관리
+ * Zustand stores and selectors
+ */
+
+// Auth Store
+export {
+  useAuthStore,
+  selectUser,
+  selectIsAuthenticated,
+  selectIsLoading,
+  selectError,
+  selectIsAdmin,
+  mockLogin,
+} from './authStore';
